bs58::encode

Trait EncodeTarget

pub trait EncodeTarget {
    // Required method
    fn encode_with(
        &mut self,
        max_len: usize,
        f: impl for<'a> FnOnce(&'a mut [u8]) -> Result<usize>,
    ) -> Result<usize>;
}

Represents a buffer that can be encoded into. See EncodeBuilder::onto and the provided implementations for more details.

Required Methods

fn encode_with( &mut self, max_len: usize, f: impl for<'a> FnOnce(&'a mut [u8]) -> Result<usize>, ) -> Result<usize>

Encodes into this buffer, provides the maximum length for implementations that wish to preallocate space, along with a function that will encode ASCII bytes into the buffer and return the length written to it.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.