Trait bs58::encode::EncodeTarget

pub trait EncodeTarget {
    fn encode_with(
        &mut self, 
        max_len: usize, 
        f: impl for<'a> FnOnce(&'a mut [u8]) -> Result<usize>
    ) -> Result<usize>;
}

Represents a buffer that can be encoded into. See EncodeBuilder::into and the provided implementations for more details.

Required methods

fn encode_with(
&mut self,
max_len: usize,
f: impl for<'a> FnOnce(&'a mut [u8]) -> Result<usize>
) -> Result<usize>

Encodes into this buffer, provides the maximum length for implementations that wish to preallocate space, along with a function that will encode ASCII bytes into the buffer and return the length written to it.